requirements.txt
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.0
Django==5.1.4
djangorestframework==3.15.2
idna==3.10
requests==2.32.3
sqlparse==0.5.3
urllib3==2.2.3
gunicorn==21.2.0
whitenoise==6.5.0
psycopg2==2.9.3
dj-database-url==2.0.0
python-decouple==3.8
pillow==10.0.0
django-environ==0.10.0
django-background-tasks==1.2.8
numpy==2.2.0
ta-lib==0.5.1